Add graceful shutdown of API when receiving SIGTERM
