--- services/api/src/schema.js
@@ -1,8 +1,0 @@
-const { makeExecutableSchema } = require('graphql-tools');
-
-const typeDefs = require('./typeDefs');
-const resolvers = require('./resolvers');
-
-module.exports = {
-  schema: makeExecutableSchema({ typeDefs, resolvers }),
-};
